Switch to plaintext for SDK test harness streaming connection (#73)
Because the Roku cannot natively handle SSE, it must interact with
alternative endpoints. These endpoints communicate via custom
encryption.

Our SDK test harness does not support this encryption. Therefore, to
fully integrate with the test harness, we need to introduce a way for
the SDK to communicate with the streaming endpoint in plain text mode.

This commit introduces a private variable on the config we can toggle to
enable this plain text mode, which we do for the contract tests only.

A related commit will be introduced into the SDK test harness to provide
support for the Roku's alternative endpoints.