fix: Fix evaluation race between status and flag fields
When the SDK is first initializing, the flag field is set before the
status field on the LaunchDarkly node. As a result, listeners on the
flag field would be executed before the status field is changed.

If a flag change listener tries to evaluate a flag the first time a
change is triggered, the default value will be returned along with an
error that the client isn't ready since the SDK doesn't think it is
initialized.

To fix this, evaluation is allowed to continue if the SDK is initialized
or if the store has flag data present. This is inline with other mobile
SDKs.